Improved error responses of bot service and form service.

+++ src/modules/form/form.controller.ts
@@ -66,6 +66,9 @@ export class FormController {
   async single(
     @UploadedFiles() files: {form: Express.Multer.File[], mediaFiles: Express.Multer.File[]},
   ) {
+    if (!files.form[0]) {
+      throw new BadRequestException('Form file is required!');
+    }
     const response = await this.formService.uploadForm(files.form[0], files.mediaFiles);
     fs.unlink(files.form[0].path, (err) => {
       if (err) {
